+++ hikyuu_cpp/hikyuu/strategy/RunSystemInStrategy.cpp
@@ -95,11 +95,11 @@ StrategyPtr HKU_API crtSysStrategy(const SYSPtr& sys, const string& stk_market_c
     };
 
     KQuery::KType ktype = query.kType();
-    StrategyPtr stg = std::make_shared<Strategy>(vector<string>{stk_market_code, "SH000001"},
-                                                 vector<KQuery::KType>{ktype},
-                                                 unordered_map<string, int>{}, name, config_file);
+    StrategyPtr stg = std::make_shared<Strategy>(
+      vector<string>{stk_market_code, "SH000001"}, vector<KQuery::KType>{ktype},
+      unordered_map<string, int64_t>{}, name, config_file);
 
-    int32_t m = KQuery::getKTypeInSeconds(ktype);
+    int64_t m = KQuery::getKTypeInSeconds(ktype);
     if (m < KQuery::getKTypeInSeconds(KQuery::DAY)) {
         stg->runDaily(std::move(func), Seconds(m), "SH");
     } else {
